Fixed a bug where the plugin was imported before setting Rider as the default scripting app
Since the plugin executes when reloading the app domain, in case Rider is not set as the default scripting app, it will not initialize itself.
Then, clicking a script file will fail, since the init was never performed.

The fix is to mark the plugin as initialized, and in case a file is being opened but the plugin isn’t initialized — force initialization.